pyproject.toml
[project]
name = "spotify-recommendation-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.9"
dependencies = [
    "pyyaml>=6.0.2",
    "spotipy>=2.25.1",
    "typer>=0.16.0",
]
